Create long annual tables from line listing data

From FY 2018/19, annual tables will be moving to a long format rather than a wide format. This function takes line-listing data and aggregates by organisation and time period, producing counts of cases by a given metric (apportioning or prior healthcare exposure).

ann_tab_long(dat, collection_var, org_var, metric_var, period_var)

Arguments

dat
A dataframe containing line-listing data

collection_var
A string giving the data collection

org_var
The variable giving the organisation code

metric_var
Column in dat giving column for the metric

period_var
Column giving the period by which counts should be aggregated

Value

A dataframe with counts of cases in long format, by organisation and time period

Details

The data are expanded to cover time periods in the data, but there may well be organisations that are not in the line-listing data that will be merged in later to make a complete table.

In addition, PHEC and organisation names will need to be merged in, as will denominator data, and rates will need to be calculated.

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

data(line_listing)
line_listing$fy_long <- fy_long(line_listing$specimen_date)

line_listing %>%
  filter(collection == "C. difficile") %>%
  ann_tab_long(dat = ., collection_var = collection,
    org_var = reporting_organisation_code,
    metric_var = prior_hc, period_var = fy_long)

#> # A tibble: 28 x 5
#>    organisation period                   metric       collection   count_cases
#>    <fct>        <chr>                    <chr>        <fct>              <dbl>
#>  1 W1A          April 2015 to March 2016 all_blank    C. difficile          11
#>  2 W1A          April 2015 to March 2016 coca         C. difficile           7
#>  3 W1A          April 2015 to March 2016 coha, ha     C. difficile           3
#>  4 W1A          April 2015 to March 2016 coia         C. difficile          10
#>  5 W1A          April 2015 to March 2016 hoha, ha     C. difficile           3
#>  6 W1A          April 2015 to March 2016 Total cases  C. difficile          36
#>  7 W1A          April 2015 to March 2016 unknown_3_mo C. difficile           2
#>  8 W1A          April 2016 to March 2017 all_blank    NA                     0
#>  9 W1A          April 2016 to March 2017 coca         C. difficile           2
#> 10 W1A          April 2016 to March 2017 coha, ha     C. difficile           1
#> # ... with 18 more rows