Replicate the PIR assignment algorithm

Provisionally assigned cases are counted as if final assignment status, i.e. a case provisionally assigned to a Trust will be reported as Trust-assigned in any report.

assignment_algorithm(
  pircasestatus,
  assignmentmethodcode,
  patientlocation,
  patientcategory,
  provisionalorganisationname,
  finalpirassignedorganisationtype
)

Arguments

pircasestatus
Text string of either 'Final assignment' or 'Provisional assignment'

assignmentmethodcode
Numeric variable giving assignment method

patientlocation
Text string giving patient location

patientcategory
Text string giving patient inpatient status (one of "", "A&E only", "Day patient", "Emergency Assessment" "In-patient", "Other", "Outpatient", "Regular Attender", "Unknown" )

provisionalorganisationname
Text string giving the name of the provisionally assigned organisation

finalpirassignedorganisationtype
Text string giving organisation type of finally assigned pir organisation. One of either "NHS Trust" or "Clinical Commissioning Group"

Value

string giving the name of the finally assigned organisation or one of "NHS Trust" or "CCG"

Examples

data(assignment_data)
assignment_data$new <- assignment_algorithm(pircasestatus = assignment_data$pircasestatus,
        assignmentmethodcode = assignment_data$assignmentmethodcode,
        patientlocation = assignment_data$patientlocation,
        patientcategory = assignment_data$patient_category,
        provisionalorganisationname = assignment_data$provisionalorganisationname,
        finalpirassignedorganisationtype = assignment_data$finalpirassignedorganisation)

#> Warning: coercing argument of type 'double' to logical

assignment_data

#>             pircasestatus assignmentmethodcode             patientlocation
#> 1        Final assignment                    9             NHS Acute Trust
#> 2        Final assignment                   10             NHS Acute Trust
#> 3        Final assignment                   10 Independent Sector Provider
#> 4        Final assignment                    8 Independent Sector Provider
#> 5  Provisional assignment                   10             NHS Acute Trust
#> 6  Provisional assignment                   10 Independent Sector Provider
#> 7  Provisional assignment                   13             NHS Acute Trust
#> 8  Provisional assignment                   15             NHS Acute Trust
#> 9  Provisional assignment                   15             NHS Acute Trust
#> 10 Provisional assignment                   11             NHS Acute Trust
#> 11 Provisional assignment                   11             NHS Acute Trust
#> 12 Provisional assignment                   11             NHS Acute Trust
#> 13 Provisional assignment                   NA             NHS Acute Trust
#> 14 Provisional assignment                   NA             NHS Acute Trust
#>                            provisionalorganisationname
#> 1  Cambridge University Hospitals NHS Foundation Trust
#> 2  Cambridge University Hospitals NHS Foundation Trust
#> 3  Cambridge University Hospitals NHS Foundation Trust
#> 4  Cambridge University Hospitals NHS Foundation Trust
#> 5  Cambridge University Hospitals NHS Foundation Trust
#> 6                               NHS Central London CCG
#> 7  Cambridge University Hospitals NHS Foundation Trust
#> 8  Cambridge University Hospitals NHS Foundation Trust
#> 9                               NHS Central London CCG
#> 10 Cambridge University Hospitals NHS Foundation Trust
#> 11                              NHS Central London CCG
#> 12                              NHS Central London CCG
#> 13                              NHS Central London CCG
#> 14                              NHS Central London CCG
#>    finalpirassignedorganisation patient_category                          new
#> 1                     NHS Trust       In-patient                    NHS Trust
#> 2                     NHS Trust       In-patient                    NHS Trust
#> 3  Clinical Commissioning Group       In-patient Clinical Commissioning Group
#> 4  Clinical Commissioning Group       In-patient Clinical Commissioning Group
#> 5                     NHS Trust       In-patient                    NHS Trust
#> 6                     NHS Trust       In-patient Clinical Commissioning Group
#> 7                     NHS Trust       In-patient                  Third Party
#> 8                     NHS Trust       In-patient                    NHS Trust
#> 9                     NHS Trust       In-patient Clinical Commissioning Group
#> 10                    NHS Trust       In-patient                    NHS Trust
#> 11                    NHS Trust         A&E only Clinical Commissioning Group
#> 12                    NHS Trust       In-patient                    NHS Trust
#> 13                    NHS Trust       In-patient Clinical Commissioning Group
#> 14                    NHS Trust       In-patient Clinical Commissioning Group

dat2 <- assignment_data[13, ]
dat2$new <- NULL
dat2$new <- assignment_algorithm(pircasestatus = dat2$pircasestatus,
        assignmentmethodcode = dat2$assignmentmethodcode,
        patientlocation = dat2$patientlocation,
        patientcategory = dat2$patient_category,
        provisionalorganisationname = dat2$provisionalorganisationname,
        finalpirassignedorganisationtype = dat2$finalpirassignedorganisation)

#> Warning: coercing argument of type 'double' to logical

dat2

#>             pircasestatus assignmentmethodcode patientlocation
#> 13 Provisional assignment                   NA NHS Acute Trust
#>    provisionalorganisationname finalpirassignedorganisation patient_category
#> 13      NHS Central London CCG                    NHS Trust       In-patient
#>                             new
#> 13 Clinical Commissioning Group

data(assignment_data2)
assignment_data2$new <- assignment_algorithm(pircasestatus = assignment_data2$pir_case_status,
        assignmentmethodcode = assignment_data2$assignment_method_code,
        patientlocation = assignment_data2$patient_location,
        patientcategory = assignment_data2$patient_category,
        provisionalorganisationname = assignment_data2$provisional_organisation_name,
        finalpirassignedorganisationtype = assignment_data2$final_pir_assigned_organisation_type)

#> Warning: coercing argument of type 'double' to logical

assignment_data2

#>           pir_case_status assignment_method_code patient_location
#> 1  Provisional Assignment                     NA  NHS Acute Trust
#> 2  Provisional Assignment                     NA  NHS Acute Trust
#> 3  Provisional Assignment                     NA  NHS Acute Trust
#> 4  Provisional Assignment                     NA  NHS Acute Trust
#> 5  Provisional Assignment                     NA  NHS Acute Trust
#> 6  Provisional Assignment                     NA  NHS Acute Trust
#> 7  Provisional Assignment                     NA  NHS Acute Trust
#> 8  Provisional Assignment                     NA  NHS Acute Trust
#> 9  Provisional Assignment                     NA  NHS Acute Trust
#> 10 Provisional Assignment                     NA  NHS Acute Trust
#> 11 Provisional Assignment                     NA  NHS Acute Trust
#> 12 Provisional Assignment                     NA  NHS Acute Trust
#> 13 Provisional Assignment                     NA  NHS Acute Trust
#> 14 Provisional Assignment                     NA  NHS Acute Trust
#> 15 Provisional Assignment                     NA  NHS Acute Trust
#>        patient_category           provisional_organisation_name
#> 1            In-patient                   NHS TOWER HAMLETS CCG
#> 2            In-patient                        NHS CHILTERN CCG
#> 3            In-patient                 NHS GLOUCESTERSHIRE CCG
#> 4  Emergency Assessment                         NHS BRISTOL CCG
#> 5            In-patient NHS DARTFORD, GRAVESHAM AND SWANLEY CCG
#> 6  Emergency Assessment                         NHS CRAWLEY CCG
#> 7            In-patient             NHS BRACKNELL AND ASCOT CCG
#> 8            In-patient                        NHS SOMERSET CCG
#> 9            In-patient           NHS SOUTH GLOUCESTERSHIRE CCG
#> 10           In-patient         NHS HIGH WEALD LEWES HAVENS CCG
#> 11 Emergency Assessment                         NHS BRISTOL CCG
#> 12           In-patient                   NHS TOWER HAMLETS CCG
#> 13           In-patient                  NHS NORTHUMBERLAND CCG
#> 14 Emergency Assessment                         NHS BRISTOL CCG
#> 15          Day patient                      NHS WEST ESSEX CCG
#>    final_pir_assigned_organisation_type id                          new
#> 1                                        1 Clinical Commissioning Group
#> 2                                        2 Clinical Commissioning Group
#> 3                                        3 Clinical Commissioning Group
#> 4                                        4 Clinical Commissioning Group
#> 5                                        5 Clinical Commissioning Group
#> 6                                        6 Clinical Commissioning Group
#> 7                                        7 Clinical Commissioning Group
#> 8                                        8 Clinical Commissioning Group
#> 9                                        9 Clinical Commissioning Group
#> 10                                      10 Clinical Commissioning Group
#> 11                                      11 Clinical Commissioning Group
#> 12                                      12 Clinical Commissioning Group
#> 13                                      13 Clinical Commissioning Group
#> 14                                      14 Clinical Commissioning Group
#> 15                                      15 Clinical Commissioning Group