Compare two data frames to identify cells which are different

Used to identify cells which have changed between publications for later highlighting in Excel export. With thanks to Amy Mikhail for writing this function for me.

data_compare(currentdf, lastdf)

Arguments

currentdf
A data frame in wide format giving the values for the current publication.

lastdf
A data frame in wide format giving the values for the most recent publication prior to the one being prepared.

Value

A matrix specifying cells that have changed.

Examples

last_pub <- data.frame(org = c("a", "b", "c"),
jan16 = sample(1:100, 3, replace=T),
feb16 = sample(1:100, 3, replace=T),
mar16 = sample(1:100, 3, replace=T),
jan17 = sample(1:100, 3, replace=T))

# copy the last pub, but drop the oldest month
this_pub <- last_pub[c(1,3:5)]
# add new month
this_pub$feb17 <- sample(1:100, 3, replace=T)
# change one value only
this_pub$mar16[2] <- this_pub$mar16[2] + 1
# Checki it out:
this_pub

#>   org feb16 mar16 jan17 feb17
#> 1   a    16    50    78     4
#> 2   b     1    30    88    33
#> 3   c    47    74    18    41

highlight <- data_compare(this_pub, last_pub)