graph_age_groups

Produces nice age groups for plotting on graphs. Different output for CDI and BSI

graph_age_groups(x, data_collection)

Arguments

x
A character vector of age groups

data_collection
a string giving the data collection of the labels to be converted.

Value

A factor with labels that can be printed in ggplot2

See also

Examples

data(age_trends_data)
head(age_trends_data)

#>   fyear6 age_group_new    sex     n       popn   rate age_sex_pc
#> 1 200708          2-14 Female   157  3858142.2    4.1  0.4920859
#> 2 200708         15-44 Female  1539 10721325.4   14.4  4.8236953
#> 3 200708         45-64 Female  3333  6463536.8   51.6 10.4466385
#> 4 200708         65-74 Female  4904  2205682.2  222.3 15.3706316
#> 5 200708         75-84 Female 10721  1655252.5  647.7 33.6028836
#> 6 200708          ge85 Female 11251   755018.8 1490.2 35.2640652

## To get the right sorting and labels for the age groups

age_trends_data$age_group2 <- graph_age_groups(age_trends_data$age_group_new,
  data_collection = "cdi")