Italicise species names for graphing Producing mixed italics for species names in plots is tricky. This function will do it for the species in the mandatory surveillance collections

Italicise species names for graphing Producing mixed italics for species names in plots is tricky. This function will do it for the species in the mandatory surveillance collections

graph_species_names(x)

Arguments

x
a character vector giving the data collection

Value

A factor variable with labels that will control italicisation

Examples

graph_species_names("Klebsiella spp")

#> [1] paste(italic("Klebsiella "), " spp.")
#> Levels: paste(italic("Klebsiella "), " spp.")

graph_species_names("MRSA")

#> [1] MRSA
#> Levels: MRSA

graph_species_names(x = c("Klebsiella spp", "MRSA"))

#> [1] paste(italic("Klebsiella "), " spp.") MRSA                                 
#> Levels: paste(italic("Klebsiella "), " spp.") MRSA

graph_species_names <- function(x){
  assertthat::assert_that(is.character(x), msg = "x must be a character")
  x <- trimws(x)
  assertthat::assert_that(
    all(x %in% c("C. difficile", "E. coli","Klebsiella spp", "MRSA", "MSSA",
                  "P. aeruginosa", "Pseudomonas aeruginosa")),
                  msg = "Please make sure that x is one of C. difficile, E. coli,
                  Klebsiella spp, MRSA, MSSA, P. aeruginosa or Pseudomonas aeruginosa,
    S. aureus or Staphylococcus aureus"
    )
  x <- ifelse(x == "Pseudomonas aeruginosa", "P. aeruginosa", x)
  x <- factor(x)
  levels(x) = c("C. difficile" = expression(italic("C. difficile")),
                "E. coli" = expression(italic("E. coli")),
                "Klebsiella spp" = expression(paste(italic("Klebsiella "), " spp.")),
                "MRSA" = "MRSA",
                "MSSA" = "MSSA",
                "P. aeruginosa" = expression(italic("P. aeruginosa")),
                "S. aureus" = expression(italic("S. aureus"))
  )
  x <- droplevels(x)
  return(x)
}