Data simulating output from the HCAI DCS line listing report

The line-listing report of the HCAI DCS provides data on cases entered onto the system. There is one row per case of infection. This dataset replicates the data structure of the line listing report for a selection of variables for use in testing and in examples.

line_listing

Format

A dataframe with 500 rows of 10 variables

id

Integer to simulate the record id

collection

A string giving the name of the organism collection

age

Patient age in years

sex

String giving patient sex

specimen_date

Date giving the date the specimen was taken

onset_status

String, whether hospital-onset or community-onset

reporting_organisation_code

A three-letter code for the reporting organisation

ccg_code

A three-letter code giving the attributed CCG

pir_status

A string giving the PIR status (only relevant to MRSA)

prior_hc

A string giving the prior healthcare group of the case (CDI only)

patient_location

A string giving location of patient at time of specimen

patient_category

A string giving admission status of patient at time of specimen

Details

All the data in the data set are simulated and do not represent actual patient data.