lag mf trend data

Creates twelve month lag, three month sum, lagged by twelve months, and twelve month sum, lagged by 12 months.

mf_lag_trend(x = trend)

Arguments

x
A data frame with columns for "cdi", "ecoli", "mrsa", "mssa", "kleb", "paer" and "ecoli_ta".

Value

The same dataframe, but with new lagged variables added

Examples

data(mf_trend_data)
library(dplyr)

#> Warning: package 'dplyr' was built under R version 3.5.1

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:lubridate':
#> 
#>     intersect, setdiff, union

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

mf_trend_data <- mf_trend_data %>% mf_lag_trend()

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

head(mf_trend_data)

#>   year month  cdi ecoli mrsa mssa kleb paer ecoli_ta          t cdi_lag_12
#> 1 2014     1  996  2878   70  803   NA   NA      644 2014-01-01         NA
#> 2 2014     2  913  2625   59  764   NA   NA      601 2014-02-01         NA
#> 3 2014     3 1098  2877   77  837   NA   NA      636 2014-03-01         NA
#> 4 2014     4 1042  2873   49  792   NA   NA      615 2014-04-01         NA
#> 5 2014     5 1214  3025   64  772   NA   NA      603 2014-05-01         NA
#> 6 2014     6 1193  2988   68  753   NA   NA      584 2014-06-01         NA
#>   ecoli_lag_12 mrsa_lag_12 mssa_lag_12 kleb_lag_12 paer_lag_12 ecoli_ta_lag_12
#> 1           NA          NA          NA          NA          NA              NA
#> 2           NA          NA          NA          NA          NA              NA
#> 3           NA          NA          NA          NA          NA              NA
#> 4           NA          NA          NA          NA          NA              NA
#> 5           NA          NA          NA          NA          NA              NA
#> 6           NA          NA          NA          NA          NA              NA
#>   cdi_sum_3 ecoli_sum_3 mrsa_sum_3 mssa_sum_3 kleb_sum_3 paer_sum_3
#> 1        NA          NA         NA         NA         NA         NA
#> 2        NA          NA         NA         NA         NA         NA
#> 3      3007        8380        206       2404         NA         NA
#> 4      3053        8375        185       2393         NA         NA
#> 5      3354        8775        190       2401         NA         NA
#> 6      3449        8886        181       2317         NA         NA
#>   ecoli_ta_sum_3 cdi_sum_3_lag_3 ecoli_sum_3_lag_3 mrsa_sum_3_lag_3
#> 1             NA              NA                NA               NA
#> 2             NA              NA                NA               NA
#> 3           1881              NA                NA               NA
#> 4           1852              NA                NA               NA
#> 5           1854              NA                NA               NA
#> 6           1802              NA                NA               NA
#>   mssa_sum_3_lag_3 kleb_sum_3_lag_3 paer_sum_3_lag_3 ecoli_ta_sum_3_lag_3
#> 1               NA               NA               NA                   NA
#> 2               NA               NA               NA                   NA
#> 3               NA               NA               NA                   NA
#> 4               NA               NA               NA                   NA
#> 5               NA               NA               NA                   NA
#> 6               NA               NA               NA                   NA
#>   cdi_sum_12 ecoli_sum_12 mrsa_sum_12 mssa_sum_12 kleb_sum_12 paer_sum_12
#> 1         NA           NA          NA          NA          NA          NA
#> 2         NA           NA          NA          NA          NA          NA
#> 3         NA           NA          NA          NA          NA          NA
#> 4         NA           NA          NA          NA          NA          NA
#> 5         NA           NA          NA          NA          NA          NA
#> 6         NA           NA          NA          NA          NA          NA
#>   ecoli_ta_sum_12 cdi_sum_12_lag_12 ecoli_sum_12_lag_12 mrsa_sum_12_lag_12
#> 1              NA                NA                  NA                 NA
#> 2              NA                NA                  NA                 NA
#> 3              NA                NA                  NA                 NA
#> 4              NA                NA                  NA                 NA
#> 5              NA                NA                  NA                 NA
#> 6              NA                NA                  NA                 NA
#>   mssa_sum_12_lag_12 kleb_sum_12_lag_12 paer_sum_12_lag_12
#> 1                 NA                 NA                 NA
#> 2                 NA                 NA                 NA
#> 3                 NA                 NA                 NA
#> 4                 NA                 NA                 NA
#> 5                 NA                 NA                 NA
#> 6                 NA                 NA                 NA
#>   ecoli_ta_sum_12_lag_12
#> 1                     NA
#> 2                     NA
#> 3                     NA
#> 4                     NA
#> 5                     NA
#> 6                     NA