Monthly CCG table

Produce the monthly wide-format table for CCGs. Takes long format data for CCGs and turns it wide with 13 months as variables. Hard coded to operate on dataframe 'ccg_dat_all', it filters the data to a given collection, e.g. "MRSA" or "E. coli" as exported in the HCAI DCS line lists.

mo_std_table(data_fm, collection, column)

Arguments

data_fm
The data frame on which to work, either ccg_dat_all or trust_dat_all

collection
A text string giving the data collection

column
A column which will be transposed

Value

A wide data.frame with counts of cases per month in columns with CCGs as rows

Examples

library(dplyr)
data(monthly_ccg_data_raw)
ccg_dat_all <- monthly_ccg_data_raw %>%
    rename(org_code = ccg_code)
out_dat <- mo_std_table(data_fm = ccg_dat_all, collection = "MRSA",
    column = "total_cases")
head(out_dat)

#>   org_code July_2016 August_2016 September_2016 October_2016 November_2016
#> 1      00C         0           0              0            2             0
#> 2      00D         1           1              0            1             1
#> 3      00J         0           0              1            0             0
#> 4      00K         0           0              1            0             0
#> 5      00L         1           0              0            1             0
#> 6      00M         0           2              0            0             0
#>   December_2016 January_2017 February_2017 March_2017 April_2017 May_2017
#> 1             0            0             0          0          0        0
#> 2             0            1             0          0          1        1
#> 3             0            2             0          1          0        1
#> 4             0            0             0          0          0        0
#> 5             0            0             2          1          0        0
#> 6             0            0             1          0          2        0
#>   June_2017 July_2017
#> 1         0         0
#> 2         0         0
#> 3         1         0
#> 4         0         0
#> 5         0         0
#> 6         0         0