Data for monthly CCG table for MRSA

A sample data set for use with the mo_ccg_table function. Is output in intermediate step in production of monthly tables and has counts of cases, by CCG and month in long format.

monthly_ccg_data_raw

Format

A data frame with 714 rows and 12 variables:

data_collection

String giving the data collection

ccg_code

String giving the ccg code

fyear6

String giving 6 character financial year

year_no

Numeric giving calendar year

month_no

Numeric giving month number

total_cases

Count of all cases for CCG and month

apportioned

Count giving apportioned cases for CCG and month

nhs_trust_pir

Count of trust-assigned PIR cases

ccg_pir

Count of CCG-assigned PIR cases

third_party_pir

Count of third party-assigned PIR cases

fmonth

Numeric giving financial year month, i.e. April = 1

month_string

Character string giving name of month