Add ordered month and calendar year to a data frame

Function to add vector of ordered month-year factor to a dataframe. To be used for preparation of monthly outputs.

ordered_m_y(data, specimen_date)

Arguments

data
A data frame

specimen_date
A specimen date, in date format, as part of the data frame

Value

A dataframe with an additional column giving and ordered month in format "January 2015", ...

See also

Examples

dat <- data.frame(spec_date = c(dmy("01/05/2015", "01/06/2015", "01/05/2015",
                                    "01/07/2015", "01/08/2015", "01/09/2015",
                                    "01/10/2015", "01/11/2015", "01/12/2015",
                                    "01/01/2016", "01/02/2016", "01/03/2016",
                                    "01/04/2016", "01/05/2016", "01/04/2015")),
                  stringsAsFactors = FALSE)
dat <- ordered_m_y(dat, "spec_date")

#> Error in mutate_impl(.data, dots): Evaluation error: factor level [3] is duplicated.

dat

#>     spec_date
#> 1  2015-05-01
#> 2  2015-06-01
#> 3  2015-05-01
#> 4  2015-07-01
#> 5  2015-08-01
#> 6  2015-09-01
#> 7  2015-10-01
#> 8  2015-11-01
#> 9  2015-12-01
#> 10 2016-01-01
#> 11 2016-02-01
#> 12 2016-03-01
#> 13 2016-04-01
#> 14 2016-05-01
#> 15 2015-04-01

head(dat$month_yr)

#> NULL