ann_tab_onset_annual

Produce annual tables by onset status

ann_tab_onset_annual(dat, timeperiod, org_code, denominator, total, ho, co,
  org_type)

Arguments

dat
A dataframe in long format

timeperiod
A financial year as YYYYYY or YYYY/YY. Can be either numeric or character

org_code
Character vector giving organisation code

denominator
The denominator value

total
Total count of cases

ho
Count of hospital-onset cases

co
Count of community-onset cases

org_type
The only parameter which should not be a column in the data frame a character giving either "Trust" or "CCG"

Value

A wide dataframe with total counts, total rates, onset count, onset rate, other onset count by financial year

Examples

ann_onset_testdat <- data.frame(stringsAsFactors=FALSE,
              time_period = c(201415L, 201516L, 201617L, 201718L, 201415L, 201516L,
                              201617L, 201718L),
                 org_code = c("W1A", "W1A", "W1A", "W1A", "E17", "E17", "E17",
                              "E17"),
              denominator = c(100L, 100L, 100L, 100L, 100L, 100L, 100L, 100L),
                    this_is_the_total = c(10L, 20L, 30L, 40L, 50L, 0L, 40L, 30L),
                       ho = c(8L, 16L, 22L, 31L, 40L, 0L, 35L, 15L),
                       co = c(2L, 4L, 8L, 9L, 10L, 0L, 5L, 15L),
       pir_trust_assigned = c(4L, 10L, 5L, 30L, 10L, 0L, 5L, 5L),
         pir_ccg_assigned = c(3L, 5L, 20L, 5L, 40L, 0L, 30L, 20L),
              third_party = c(3L, 5L, 5L, 5L, 0L, 0L, 5L, 5L)
    )
 ann_tab_onset_annual(ann_onset_testdat, timeperiod = time_period,
    org_code = org_code, denominator = denominator, total= this_is_the_total, ho = ho,
    co = co, org_type = "trust")

#>   org_code April 2014 to March 2015_Total cases*
#> 1      E17                                    50
#> 2      W1A                                    10
#>   April 2014 to March 2015_Total rate April 2014 to March 2015_HO cases**
#> 1                               50000                                  40
#> 2                               10000                                   8
#>   April 2014 to March 2015_HO rate April 2014 to March 2015_CO cases***
#> 1                            40000                                   10
#> 2                             8000                                    2
#>   April 2015 to March 2016_Total cases* April 2015 to March 2016_Total rate
#> 1                                     0                                   0
#> 2                                    20                               20000
#>   April 2015 to March 2016_HO cases** April 2015 to March 2016_HO rate
#> 1                                   0                                0
#> 2                                  16                            16000
#>   April 2015 to March 2016_CO cases*** April 2016 to March 2017_Total cases*
#> 1                                    0                                    40
#> 2                                    4                                    30
#>   April 2016 to March 2017_Total rate April 2016 to March 2017_HO cases**
#> 1                               40000                                  35
#> 2                               30000                                  22
#>   April 2016 to March 2017_HO rate April 2016 to March 2017_CO cases***
#> 1                            35000                                    5
#> 2                            22000                                    8
#>   April 2017 to March 2018_Total cases* April 2017 to March 2018_Total rate
#> 1                                    30                               30000
#> 2                                    40                               40000
#>   April 2017 to March 2018_HO cases** April 2017 to March 2018_HO rate
#> 1                                  15                            15000
#> 2                                  31                            31000
#>   April 2017 to March 2018_CO cases***
#> 1                                   15
#> 2                                    9