ann_tab_onset_qtrly

Prepare quarterly onset tables. Takes data in long format and reshapes wide for use in annual tables.

ann_tab_onset_qtrly(dat, timeperiod, org_code, total, ho, co, org_type)

Arguments

dat
A data frame with the necessary columns

timeperiod
A financial year quarter

org_code
Character vector giving organisation code

total
Total count of cases

ho
Count of hospital-onset cases

co
Count of community-onset cases

org_type
The only parameter which should not be a column in the data frame a character giving either "Trust" or "CCG"

Value

A wide dataframe with total counts, total rates, onset count, onset rate, other onset count by financial year

See also

Examples

ann_onset_testdat <- data.frame(stringsAsFactors=FALSE,
              time_period = c(20141L, 20142L, 20153L, 20164L, 20141L, 20152L,
                              20163L, 20174L),
                 org_code = c("W1A", "W1A", "W1A", "W1A", "E17", "E17", "E17",
                              "E17"),
              denominator = c(100L, 100L, 100L, 100L, 100L, 100L, 100L, 100L),
                    this_is_the_total = c(10L, 20L, 30L, 40L, 50L, 0L, 40L, 30L),
                       ho = c(8L, 16L, 22L, 31L, 40L, 0L, 35L, 15L),
                       co = c(2L, 4L, 8L, 9L, 10L, 0L, 5L, 15L),
       pir_trust_assigned = c(4L, 10L, 5L, 30L, 10L, 0L, 5L, 5L),
         pir_ccg_assigned = c(3L, 5L, 20L, 5L, 40L, 0L, 30L, 20L),
              third_party = c(3L, 5L, 5L, 5L, 0L, 0L, 5L, 5L)
    )
 ann_tab_onset_qtrly(ann_onset_testdat, timeperiod = time_period,
    org_code = org_code, total= this_is_the_total, ho = ho,
    co = co, org_type = "trust")

#>   org_code April to June 2014_Total cases April to June 2014_HO cases
#> 1      E17                             50                          40
#> 2      W1A                             10                           8
#>   April to June 2014_CO cases July to September 2014_Total cases
#> 1                          10                                  0
#> 2                           2                                 20
#>   July to September 2014_HO cases July to September 2014_CO cases
#> 1                               0                               0
#> 2                              16                               4
#>   July to September 2015_Total cases July to September 2015_HO cases
#> 1                                  0                               0
#> 2                                  0                               0
#>   July to September 2015_CO cases October to December 2015_Total cases
#> 1                               0                                    0
#> 2                               0                                   30
#>   October to December 2015_HO cases October to December 2015_CO cases
#> 1                                 0                                 0
#> 2                                22                                 8
#>   October to December 2016_Total cases October to December 2016_HO cases
#> 1                                   40                                35
#> 2                                    0                                 0
#>   October to December 2016_CO cases January to March 2017_Total cases
#> 1                                 5                                 0
#> 2                                 0                                40
#>   January to March 2017_HO cases January to March 2017_CO cases
#> 1                              0                              0
#> 2                             31                              9
#>   January to March 2018_Total cases January to March 2018_HO cases
#> 1                                30                             15
#> 2                                 0                              0
#>   January to March 2018_CO cases
#> 1                             15
#> 2                              0