Data for testing prior healthcare exposure monthly table

A sample data set to test the monthly table preparation for prior healthcare exposure

cdi_prior_hc_data

Format

A data frame with 13 rows of 16 variables

org_code

Organisation code (CCG or NHS trust)

year_no

Four-digit year

month_no

Numeric month

data_collection

Data collection

fyear6

Financial year

total_cases

Count of total cases

apportioned

Count of apportioned cases

nhs_trust_pir

Count of trust-assigned cases

ccg_pir

Count of ccg-assigned cases

third_party_pir

Count of third-party cases

hoha

Count of hoha cases

coha

Count of coha cases

coia

Count of coia cases

coca

Count of coca cases

phc_all_blank

Count of cases for which all prior health care fields are all blank

phc_unknown_3_mo

Count of cases for which the three-month prior health care fields is blank

fmonth

Numeric financial year month (i.e. 1 = April)

month_string

String cal month