Monthly table for Gram-negative blood stream infection data

Produce the monthly wide-format table for GNBSI . Takes long format data for either trusts or CCGs and turns it wide with 13 months as variables. The function will filter the data to a given collection, i.e. "K. pneumoniae", "E. coli" or "Pseudomonas" as exported in the HCAI DCS line lists.

mo_gnabsi_table(data_fm, collection, org_type)

Arguments

data_fm
A data frame to work on, usually either trust_dat_all or ccg_dat_all

collection
A text string giving the data collection

org_type
A string giving the organisation type, one of trust or ccg. Lower case.

Value

A wide data.frame with counts of cases by org and month.

Examples

library(dplyr)
data(monthly_ccg_data_raw)
head(monthly_ccg_data_raw)

#>   data_collection ccg_code fyear6 year_no month_no total_cases apportioned
#> 1            MRSA      00C 201617    2016       10           2           1
#> 2            MRSA      00D 201617    2016        7           1           1
#> 3            MRSA      00D 201617    2016        8           1           0
#> 4            MRSA      00D 201617    2016       10           1           1
#> 5            MRSA      00D 201617    2016       11           1           1
#> 6            MRSA      00D 201617    2017        1           1           1
#>   nhs_trust_pir ccg_pir third_party_pir fmonth month_string
#> 1             1       1               0      7      October
#> 2             1       0               0      4         July
#> 3             0       1               0      5       August
#> 4             1       0               0      7      October
#> 5             1       0               0      8     November
#> 6             1       0               0     12      January

ccg_dat_all <- monthly_ccg_data_raw %>%
    rename(org_code = ccg_code, ho = apportioned) %>%
    mutate(data_collection = "E. coli", co = total_cases - ho)

my_out_dat <- mo_gnabsi_table(data_fm = ccg_dat_all, collection = "E. coli",
    org_type = "ccg")

head(my_out_dat)

#>   org_code July_2016_total_cases July_2016_co July_2016_ho
#> 1      00C                     0            0            0
#> 2      00D                     1            0            1
#> 3      00J                     0            0            0
#> 4      00K                     0            0            0
#> 5      00L                     1            1            0
#> 6      00M                     0            0            0
#>   August_2016_total_cases August_2016_co August_2016_ho
#> 1                       0              0              0
#> 2                       1              1              0
#> 3                       0              0              0
#> 4                       0              0              0
#> 5                       0              0              0
#> 6                       2              0              2
#>   September_2016_total_cases September_2016_co September_2016_ho
#> 1                          0                 0                 0
#> 2                          0                 0                 0
#> 3                          1                 0                 1
#> 4                          1                 0                 1
#> 5                          0                 0                 0
#> 6                          0                 0                 0
#>   October_2016_total_cases October_2016_co October_2016_ho
#> 1                        2               1               1
#> 2                        1               0               1
#> 3                        0               0               0
#> 4                        0               0               0
#> 5                        1               1               0
#> 6                        0               0               0
#>   November_2016_total_cases November_2016_co November_2016_ho
#> 1                         0                0                0
#> 2                         1                0                1
#> 3                         0                0                0
#> 4                         0                0                0
#> 5                         0                0                0
#> 6                         0                0                0
#>   December_2016_total_cases December_2016_co December_2016_ho
#> 1                         0                0                0
#> 2                         0                0                0
#> 3                         0                0                0
#> 4                         0                0                0
#> 5                         0                0                0
#> 6                         0                0                0
#>   January_2017_total_cases January_2017_co January_2017_ho
#> 1                        0               0               0
#> 2                        1               0               1
#> 3                        2               0               2
#> 4                        0               0               0
#> 5                        0               0               0
#> 6                        0               0               0
#>   February_2017_total_cases February_2017_co February_2017_ho
#> 1                         0                0                0
#> 2                         0                0                0
#> 3                         0                0                0
#> 4                         0                0                0
#> 5                         2                1                1
#> 6                         1                1                0
#>   March_2017_total_cases March_2017_co March_2017_ho April_2017_total_cases
#> 1                      0             0             0                      0
#> 2                      0             0             0                      1
#> 3                      1             1             0                      0
#> 4                      0             0             0                      0
#> 5                      1             1             0                      0
#> 6                      0             0             0                      2
#>   April_2017_co April_2017_ho May_2017_total_cases May_2017_co May_2017_ho
#> 1             0             0                    0           0           0
#> 2             1             0                    1           1           0
#> 3             0             0                    1           0           1
#> 4             0             0                    0           0           0
#> 5             0             0                    0           0           0
#> 6             1             1                    0           0           0
#>   June_2017_total_cases June_2017_co June_2017_ho July_2017_total_cases
#> 1                     0            0            0                     0
#> 2                     0            0            0                     0
#> 3                     1            1            0                     0
#> 4                     0            0            0                     0
#> 5                     0            0            0                     0
#> 6                     0            0            0                     0
#>   July_2017_co July_2017_ho
#> 1            0            0
#> 2            0            0
#> 3            0            0
#> 4            0            0
#> 5            0            0
#> 6            0            0